Separate bookstore functionality out into new file
